requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==9.5.0
reportlab==4.0.4
pandas==2.1.4
openpyxl==3.1.2
python-pptx==0.6.21
PyMuPDF==1.23.8
flask-sqlalchemy==3.0.5
python-docx==0.8.11 